+++ CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [2.24.2] - Unreleased
+### Added
+  - Aggiunto job schedulato a cadenza settimanale (il lunedì) che controlla quanti dipendenti, nella settimana appena conclusa, hanno
+    superato il limite di 48 ore di lavoro settimanale previsto dalla circolare 12/2004. Per quei dipendenti viene inviata una mail
+    al responsabile della sede e all'amministratore del personale della sede di afferenza del dipendente.
+
+
 ## [2.24.1] - 2025-04-24
 ### Added
   - Gestiti timezone diversi da Europe/Rome nella creazione delle missioni via REST.